requirements.txt
asgiref==3.3.4
bcrypt==3.2.0
certifi==2020.12.5
cffi==1.14.5
Django==3.2.2
django-cors-headers==3.7.0
django-db-logger==0.1.10
djangorestframework==3.12.4
et-xmlfile==1.1.0
gunicorn==20.1.0
haversine==2.3.1
mysqlclient==2.0.3
openpyxl==3.0.7
pycparser==2.20
PyJWT==2.1.0
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
